Added new setting: SOCIAL_AUTH_NEW_USER_CHECKER.
NEW_USER_CHECKER is a function (always returning False by default) that
takes a User instance and returns True in case this user must be marked as new.
It is useful, for example, if you need to send a user to the new-users-page
each time they log in until they perform certain actions with their account.